+++ src/index.ts
@@ -1,9 +1,14 @@
 export { Sigil, SigilError } from './classes';
 export { WithSigil } from './decorator';
 export { withSigil } from './hof';
-export { isSigilCtor, isSigilInstance } from './helpers';
+export { isSigilCtor, isSigilInstance, getSigilLabels } from './helpers';
 export { Sigilify, SigilifyAbstract } from './mixin';
-export { updateSigilOptions, DEFAULT_LABEL_REGEX, type SigilOptions } from './options';
+export {
+  updateSigilOptions,
+  DEFAULT_LABEL_REGEX,
+  RECOMMENDED_LABEL_REGEX,
+  type SigilOptions,
+} from './options';
 export type {
   ISigilInstance,
   ISigilStatic,